Added PreserveAttribute to Core
