--- MvvmCross.Plugins/All/Preserve.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace MvvmCross.Plugins
-{
-	[AttributeUsage(AttributeTargets.Assembly | AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Enum | AttributeTargets.Constructor | AttributeTargets.Method | AttributeTargets.Property | AttributeTargets.Field | AttributeTargets.Event | AttributeTargets.Interface | AttributeTargets.Delegate)]
-	public sealed class PreserveAttribute : Attribute
-	{
-		public bool AllMembers;
-
-		public bool Conditional;
-	}
-}
