+++ MvvmCross/Core/Platform/PreserveAttribute.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace MvvmCross.Core.Platform
+{
+    [AttributeUsage(AttributeTargets.Assembly | AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Enum | AttributeTargets.Constructor | AttributeTargets.Method | AttributeTargets.Property | AttributeTargets.Field | AttributeTargets.Event | AttributeTargets.Interface | AttributeTargets.Delegate)]
+    public sealed class PreserveAttribute : Attribute
+    {
+        public bool AllMembers;
+        public bool Conditional;
+    }
+}
